agregada_funcion_para_tomar_klm

+++ tp_1/src/Utn_TomaDeDatos.c
@@ -53,7 +53,7 @@
  /// @param mensaje_de_Error
  /// @param contadorDeIntentos
  /// @return
-float tomarUnFlotante(float *punteroNumeroIngresado, float maximo,float minimo,char * mensajeLimites,char * mensajeError ,int contadorDeIntentos){
+int tomarUnFlotante(float *punteroNumeroIngresado, float maximo,float minimo,char * mensajeLimites,char * mensajeError ,int contadorDeIntentos){
 
 		float numeroDeSeguridad;
 		int retorno;
